refactor(webapp): split the components, allow to change loading for bp logo and fix warning
