+++ webapp/src/components/PageTitle/index.js
@@ -4,13 +4,14 @@ import PropTypes from 'prop-types'
 
 import { generalConfig } from '../../config'
 
-const PageTitle = ({ title, metaTitle, metaDescription }) => (
+const PageTitle = ({ title, metaTitle, metaDescription, ldJson }) => (
   <Helmet>
     <title>{title}</title>
     <meta name="title" content={metaTitle} />
     <meta name="description" content={metaDescription} />
     <meta property="og:title" content={metaTitle} />
     <meta property="og:description" content={metaDescription} />
+    {ldJson && <script type="application/ld+json">{ldJson}</script>}
   </Helmet>
 )
 
@@ -24,6 +25,7 @@ PageTitle.defaultProps = {
   title: generalConfig.title,
   metaTitle: '',
   metaDescription: '',
+  ldJson: null,
 }
 
 export default PageTitle
