+++ webapp/src/components/ProducerAvatar/index.js
@@ -5,13 +5,13 @@ import { generalConfig } from '../../config'
 import { onImgError } from '../../utils'
 import useBPLogoState from 'hooks/customHooks/useBPLogoState'
 
-const ProducerAvatar = ({ logo, name, classes }) => {
+const ProducerAvatar = ({ logo, name, lazy, classes }) => {
   const defaultLogo = generalConfig.defaultProducerLogo
   const [{ src, logoRef }, { handleLoad }] = useBPLogoState(logo, defaultLogo)
 
   return (
     <img
-      loading="lazy"
+      loading={lazy ? 'lazy' : 'eager'}
       className={classes?.avatar}
       src={src}
       ref={logoRef}
@@ -25,10 +25,12 @@ const ProducerAvatar = ({ logo, name, classes }) => {
 ProducerAvatar.propTypes = {
   logo: PropTypes.string,
   name: PropTypes.string,
+  lazy: PropTypes.bool,
   classes: PropTypes.object,
 }
 
 ProducerAvatar.defaultProps = {
+  lazy: true,
   clases: {},
 }
