+++ webapp/src/routes/ProducerProfile/styles.js
@@ -105,37 +105,4 @@ export default (theme) => ({
       display: 'flex',
     },
   },
-  emptyState: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    width: '100%',
-    '& span': {
-      width: '16em',
-      height: '45px',
-      fontSize: '1.2em',
-      fontWeight: 'bold',
-      fontStretch: 'normal',
-      fontStyle: 'normal',
-      lineHeight: '1.12',
-      letterSpacing: '-0.22px',
-      textAlign: 'center',
-      color: '#3d3d3dde',
-    },
-    '& a': {
-      color: theme.palette.primary.main,
-      textDecorationColor: theme.palette.primary.main,
-    },
-  },
-  imgError: {
-    [theme.breakpoints.down('lg')]: {
-      width: '200px',
-      height: '120px',
-    },
-    [theme.breakpoints.up('lg')]: {
-      width: '260px',
-      height: '160px',
-    },
-    objectFit: 'contain',
-  },
 })
